Save_14

+++ src/Drugstore/components/Azitronite.java
@@ -4,6 +4,6 @@ import Drugstore.Component;
 
 public class Azitronite extends Component {
     public Azitronite(String title, Double weight, int power) {
-        super(title, weight, (double) power);
+        super(title, weight, power);
     }
 }
