+++ src/Drugstore/components/Penicillin.java
@@ -4,6 +4,6 @@ import Drugstore.Component;
 
 public class Penicillin extends Component {
     public Penicillin(String title, Double weight, int power) {
-        super(title, weight, (double) power);
+        super(title, weight, power);
     }
 }
