+++ src/Drugstore/components/Water.java
@@ -4,6 +4,6 @@ import Drugstore.Component;
 
 public class Water extends Component {
     public Water(String title, Double weight, int power) {
-        super(title, weight, (double) power);
+        super(title, weight, power);
     }
 }
